Augusta NLCD 2011

A real landscape of area near Augusta, Georgia obtained from the National Land Cover Database (NLCD)

augusta_nlcd

Format

A raster layer object.

Source

https://www.mrlc.gov/nlcd2011.php

References

Homer, C.G., Dewitz, J.A., Yang, L., Jin, S., Danielson, P., Xian, G., Coulston, J., Herold, N.D., Wickham, J.D., and Megown, K., 2015, Completion of the 2011 National Land Cover Database for the conterminous United States-Representing a decade of land cover change information. Photogrammetric Engineering and Remote Sensing, v. 81, no. 5, p. 345-354