Calculate correlation

Calculate correlation

calculate_correlation(
  metrics,
  method = "pearson",
  diag = TRUE,
  simplify = FALSE
)

Arguments

metrics
Tibble with results of as returned by the landscapemetrics package.

method
Type of correlation. See link{cor} for details.

diag
If FALSE, values on the diagonal will be NA.

simplify
If TRUE and only one level is present, only a tibble is returned.

Value

list

Details

The functions calculates the correlation between all metrics. In order to calculate correlations, for the landscape level more than one landscape needs to be present. All input must be structured as returned by the landscapemetrics package.

Examples

metrics <- calculate_lsm(landscape, what = c("patch", "class"))

#> Warning: Please use 'check_landscape()' to ensure the input data is valid.

#> Warning: Class 1: PAFRAC = NA for class with < 10 patches

#> Warning: Class 3: PAFRAC = NA for class with < 10 patches

calculate_correlation(metrics, method = "pearson")

#> $patch
#> # A tibble: 78 x 3
#>    metric_1 metric_2  value
#>    <chr>    <chr>     <dbl>
#>  1 area     area      1    
#>  2 cai      area      0.705
#>  3 circle   area      0.415
#>  4 contig   area      0.529
#>  5 core     area      0.999
#>  6 enn      area     -0.180
#>  7 frac     area      0.298
#>  8 gyrate   area      0.909
#>  9 ncore    area      0.960
#> 10 para     area     -0.491
#> # … with 68 more rows
#> 
#> $class
#> # A tibble: 1,540 x 3
#>    metric_1  metric_2  value
#>    <chr>     <chr>     <dbl>
#>  1 ai        ai        1    
#>  2 area_cv   ai       -0.125
#>  3 area_mn   ai        0.996
#>  4 area_sd   ai        1.00 
#>  5 ca        ai        0.953
#>  6 cai_cv    ai       -0.422
#>  7 cai_mn    ai        0.780
#>  8 cai_sd    ai        0.942
#>  9 circle_cv ai       -0.942
#> 10 circle_mn ai        0.920
#> # … with 1,530 more rows
#>