get_circumscribingcircle

Diameter of the circumscribing circle around patches

get_circumscribingcircle(landscape, directions, level)

# S3 method for RasterLayer
get_circumscribingcircle(landscape, directions = 8, level = "patch")

# S3 method for RasterStack
get_circumscribingcircle(landscape, directions = 8, level = "patch")

# S3 method for RasterBrick
get_circumscribingcircle(landscape, directions = 8, level = "patch")

# S3 method for stars
get_circumscribingcircle(landscape, directions = 8, level = "patch")

# S3 method for list
get_circumscribingcircle(landscape, directions = 8, level = "patch")

Arguments

landscape
RasterLayer or matrix (with x, y, id columns)

directions
The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

level
Either 'patch' or 'class' for the corresponding level.

Details

The diameter of the smallest circumscribing circle around a patch in the landscape is based on the maximum distance between the corners of each cell. This ensures that all cells of the patch are included in the patch.

References

Based on C++ code from Project Nayuki (https://www.nayuki.io/page/smallest-enclosing-circle).

Examples

# get circle around each patch
get_circumscribingcircle(landscape)

#> # A tibble: 27 x 7
#>    layer level class    id value center_x center_y
#>    <int> <chr> <int> <int> <dbl>    <dbl>    <dbl>
#>  1     1 patch     1     1  1.41      0.5     29.5
#>  2     1 patch     1     2  4.12      0.5     21  
#>  3     1 patch     1     3 29.3      13.2     19.2
#>  4     1 patch     1     4  1.41      0.5      2.5
#>  5     1 patch     1     5  1.41      0.5      0.5
#>  6     1 patch     1     6  6.32     17        8  
#>  7     1 patch     1     7  3.16     20.5      0.5
#>  8     1 patch     1     8  3.61     26        2.5
#>  9     1 patch     1     9  1.41     29.5      6.5
#> 10     1 patch     2    10 10.3       3.5     25.5
#> # … with 17 more rows

# get circle around whole class
get_circumscribingcircle(landscape, level = "class")

#> # A tibble: 3 x 7
#>   layer level class    id value center_x center_y
#>   <int> <chr> <int> <int> <dbl>    <dbl>    <dbl>
#> 1     1 class     1    NA  39.2     12.6     15  
#> 2     1 class     2    NA  41.7     15       14.5
#> 3     1 class     3    NA  41.7     15.2     15.2