landscapemetrics

Calculates landscape metrics for categorical landscape patterns in a tidy workflow. 'landscapemetrics' reimplements the most common metrics from 'FRAGSTATS' (https://www.fragstats.org/) and adds new ones from the current literature on landscape metrics. This package supports 'terra' SpatRaster objects as input arguments. It further provides utility functions to visualize patches, select metrics and building blocks to develop new metrics.

See also

Author

Maintainer: Maximilian H.K. Hesselbarth mhk.hesselbarth@gmail.com (ORCID)

Authors:

Marco Sciaini sciaini.marco@gmail.com (ORCID)

Jakub Nowosad nowosad.jakub@gmail.com (ORCID)

Sebastian Hanss (ORCID)

Other contributors:

Laura J. Graham (Input on package structure) [contributor]

Jeffrey Hollister (Input on package structure) [contributor]

Kimberly A. With (Input on package structure) [contributor]

Florian Privé (Original author of underlying C++ code for get_nearestneighbour() function) [contributor]

Project Nayuki (Original author of underlying C++ code for get_circumscribingcircle and lsm_p_circle) [contributor]

Matt Strimas-Mackey (Bugfix in sample_metrics()) [contributor]