List landscape metrics

List landscape metrics

list_lsm(level = NULL, metric = NULL, name = NULL, type = NULL,
  what = NULL, simplify = FALSE, verbose = TRUE)

Arguments

level
Level of metrics. Either 'patch', 'class' or 'landscape' (or vector with combination).

metric
Abbreviation of metrics (e.g. 'area').

name
Full name of metrics (e.g. 'core area')

type
Type according to FRAGSTATS grouping (e.g. 'aggregation metrics').

what
Selected level of metrics: either "patch", "class" or "landscape". It is also possible to specify functions as a vector of strings, e.g. what = c("lsm_c_ca", "lsm_l_ta").

simplify
If true, function names are returned as vector.

verbose
Print warning messages

Value

tibble

Details

List all available landscape metrics depending on the provided filter arguments. If an argument is not provided, automatically all possibilities are selected. Furthermore, it is possible to only get a vector with all function names instead of the full tibble.

References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Examples

list_lsm(level = c("patch", "landscape"), type = "aggregation metric")

#> # A tibble: 15 x 5
#>    metric   name                          type            level   function_name
#>    <chr>    <chr>                         <chr>           <chr>   <chr>        
#>  1 enn      euclidean nearest neighbor d… aggregation me… patch   lsm_p_enn    
#>  2 ai       aggregation index             aggregation me… landsc… lsm_l_ai     
#>  3 cohesion patch cohesion index          aggregation me… landsc… lsm_l_cohesi…
#>  4 contag   connectance                   aggregation me… landsc… lsm_l_contag 
#>  5 division division index                aggregation me… landsc… lsm_l_divisi…
#>  6 enn_cv   euclidean nearest neighbor d… aggregation me… landsc… lsm_l_enn_cv 
#>  7 enn_mn   euclidean nearest neighbor d… aggregation me… landsc… lsm_l_enn_mn 
#>  8 enn_sd   euclidean nearest neighbor d… aggregation me… landsc… lsm_l_enn_sd 
#>  9 iji      interspersion and juxtaposit… aggregation me… landsc… lsm_l_iji    
#> 10 lsi      landscape shape index         aggregation me… landsc… lsm_l_lsi    
#> 11 mesh     effective mesh size           aggregation me… landsc… lsm_l_mesh   
#> 12 np       number of patches             aggregation me… landsc… lsm_l_np     
#> 13 pd       patch density                 aggregation me… landsc… lsm_l_pd     
#> 14 pladj    percentage of like adjacenci… aggregation me… landsc… lsm_l_pladj  
#> 15 split    splitting index               aggregation me… landsc… lsm_l_split  

list_lsm(metric = "area", simplify = TRUE)

#> [1] "lsm_p_area"    "lsm_c_area_cv" "lsm_c_area_mn" "lsm_c_area_sd"
#> [5] "lsm_l_area_cv" "lsm_l_area_mn" "lsm_l_area_sd"

list_lsm(metric = "area", what = "lsm_p_shape")

#> Warning: only using 'what' argument

#> # A tibble: 1 x 5
#>   metric name        type         level function_name
#>   <chr>  <chr>       <chr>        <chr> <chr>        
#> 1 shape  shape index shape metric patch lsm_p_shape  

list_lsm(metric = "area", what = c("patch", "lsm_l_ta"))

#> Warning: only using 'what' argument

#> # A tibble: 13 x 5
#>    metric name                          type             level    function_name
#>    <chr>  <chr>                         <chr>            <chr>    <chr>        
#>  1 area   patch area                    area and edge m… patch    lsm_p_area   
#>  2 cai    core area index               core area metric patch    lsm_p_cai    
#>  3 circle related circumscribing circle shape metric     patch    lsm_p_circle 
#>  4 contig contiguity index              shape metric     patch    lsm_p_contig 
#>  5 core   core area                     core area metric patch    lsm_p_core   
#>  6 enn    euclidean nearest neighbor d… aggregation met… patch    lsm_p_enn    
#>  7 frac   fractal dimension index       shape metric     patch    lsm_p_frac   
#>  8 gyrate radius of gyration            area and edge m… patch    lsm_p_gyrate 
#>  9 ncore  number of core areas          core area metric patch    lsm_p_ncore  
#> 10 para   perimeter-area ratio          shape metric     patch    lsm_p_para   
#> 11 perim  patch perimeter               area and edge m… patch    lsm_p_perim  
#> 12 shape  shape index                   shape metric     patch    lsm_p_shape  
#> 13 ta     total area                    area and edge m… landsca… lsm_l_ta     

list_lsm(what = c("lsm_c_tca", "lsm_l_ta"))

#> # A tibble: 2 x 5
#>   metric name            type                 level     function_name
#>   <chr>  <chr>           <chr>                <chr>     <chr>        
#> 1 tca    total core area core area metric     class     lsm_c_tca    
#> 2 ta     total area      area and edge metric landscape lsm_l_ta