Tibble of abbreviations coming from FRAGSTATS

A single tibble for every abbreviation of every metric that is reimplemented in landscapemetrics and its corresponding full name in the literature.

lsm_abbreviations_names

Format

A tibble object.

Details

Can be used after calculating the metric(s) with a join to have a more readable results tibble or for visualizing your results.

Examples

aggregation_index <- lsm_p_area(landscape)
aggregation_index <- dplyr::right_join(aggregation_index,
                                       lsm_abbreviations_names,
                                       by = 'metric')
aggregation_index

#> # A tibble: 158 x 10
#>    layer level.x class    id metric  value name   type    level.y function_name
#>    <int> <chr>   <int> <int> <chr>   <dbl> <chr>  <chr>   <chr>   <chr>        
#>  1     1 patch       1     1 area   0.0001 patch… area a… patch   lsm_p_area   
#>  2     1 patch       1     2 area   0.0148 patch… area a… patch   lsm_p_area   
#>  3     1 patch       1     3 area   0.0005 patch… area a… patch   lsm_p_area   
#>  4     1 patch       1     4 area   0.0014 patch… area a… patch   lsm_p_area   
#>  5     1 patch       1     5 area   0.0001 patch… area a… patch   lsm_p_area   
#>  6     1 patch       1     6 area   0.0005 patch… area a… patch   lsm_p_area   
#>  7     1 patch       1     7 area   0.0001 patch… area a… patch   lsm_p_area   
#>  8     1 patch       1     8 area   0.0001 patch… area a… patch   lsm_p_area   
#>  9     1 patch       1     9 area   0.0003 patch… area a… patch   lsm_p_area   
#> 10     1 patch       2    10 area   0.0035 patch… area a… patch   lsm_p_area   
#> # ... with 148 more rows