ED (class level)

Edge Density (Area and Edge metric)

lsm_c_ed(landscape, count_boundary, directions)

# S3 method for RasterLayer
lsm_c_ed(landscape, count_boundary = FALSE,
  directions = 8)

# S3 method for RasterStack
lsm_c_ed(landscape, count_boundary = FALSE,
  directions = 8)

# S3 method for RasterBrick
lsm_c_ed(landscape, count_boundary = FALSE,
  directions = 8)

# S3 method for stars
lsm_c_ed(landscape, count_boundary = FALSE,
  directions = 8)

# S3 method for list
lsm_c_ed(landscape, count_boundary = FALSE,
  directions = 8)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

count_boundary
Count landscape boundary as edge.

directions
The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$ED = \frac{\sum \limits_{k = 1}^{m} e_{ik}} {A} * 10000$$ where \(e_{ik}\) is the total edge length in meters and \(A\) is the total landcape area in square meters.

ED is an 'Area and Edge metric'. The edge density equals the sum of all edges of class i in relation to the landcape area. The boundary of the landscape is only included in the corresponding total class edge length if count_boundary = TRUE. The metric describes the configuration of the landscape, e.g. because an aggregation of the same class will result in a low edge density. The metric is standardized to the total landscape area, and therefore comparisons among landscapes with different total areas are possible.

Units
Meters per hectare
Range
ED >= 0
Behaviour
Equals ED = 0 if only one patch is present (and the landscape boundary is not included) and increases, without limit, as the landscapes becomes more patchy
References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

See also

Examples

lsm_c_ed(landscape)

#> [38;5;246m# A tibble: 3 x 6[39m
#>   layer level class    id metric value
#>   [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m  [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m     1 class     1    [31mNA[39m ed     [4m2[24m000 
#> [38;5;250m2[39m     1 class     2    [31mNA[39m ed     [4m2[24m522.
#> [38;5;250m3[39m     1 class     3    [31mNA[39m ed     [4m3[24m567.