Fractal dimension index distribution (class level)

Coeffiecent of variation fractal dimension index (class level)

lsm_c_frac_cv(landscape)

# S3 method for RasterLayer
lsm_c_frac_cv(landscape)

# S3 method for RasterStack
lsm_c_frac_cv(landscape)

# S3 method for RasterBrick
lsm_c_frac_cv(landscape)

# S3 method for list
lsm_c_frac_cv(landscape)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

Value

tibble

Details

Coeffiecent of variation of the fractal dimension index of class i. FRAC equals two times the natural logarithm of a quarter of the patch area divided by the natural logarithm of the patch area $$FRAC_CV = cv(FRAC[patch_i])$$

Units
None
Range
???
Behaviour
???
References

McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep. PNW-351.

Examples

lsm_c_frac_cv(landscape)

#> # A tibble: 3 x 6
#>   layer level class    id metric                       value
#>   <int> <chr> <int> <int> <chr>                        <dbl>
#> 1     1 patch     1    NA fractal dimension index (cv) 14.1 
#> 2     1 patch     2    NA fractal dimension index (cv) 20.0 
#> 3     1 patch     3    NA fractal dimension index (cv)  8.58