MESH (class level)

Effective Mesh Size (Aggregation metric)

Usage

lsm_c_mesh(landscape, directions = 8)

Arguments

landscape

Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$MESH = \frac{\sum \limits_{j = 1}^{n} a_{ij} ^ 2} {A} * \frac{1} {10000}$$ where \(a_{ij}\) is the patch area in square meters and \(A\) is the total landscape area in square meters.

The effective mesh size is an 'Aggregation metric'. Because each patch is squared before the sums for each group i are calculated and the sum is standardized by the total landscape area, MESH is a relative measure of patch structure. MESH is perfectly, negatively correlated to lsm_c_division.

Units

Hectares

Range

cell size / total area <= MESH <= total area

Behaviour

Equals cellsize/total area if class covers only one cell and equals total area if only one patch is present.

References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Jaeger, J. A. 2000. Landscape division, splitting index, and effective mesh size: new measures of landscape fragmentation. Landscape ecology, 15(2), 115-130.

See also

Examples

lsm_c_mesh(landscape)
#> # A tibble: 3 × 6
#>   layer level class    id metric   value
#>   <int> <chr> <int> <int> <chr>    <dbl>
#> 1     1 class     1    NA mesh   0.00246
#> 2     1 class     2    NA mesh   0.00164
#> 3     1 class     3    NA mesh   0.0232