PLADJ (class level)

Percentage of Like Adjacencies (Aggregation metric)

lsm_c_pladj(landscape)

# S3 method for RasterLayer
lsm_c_pladj(landscape)

# S3 method for RasterStack
lsm_c_pladj(landscape)

# S3 method for RasterBrick
lsm_c_pladj(landscape)

# S3 method for stars
lsm_c_pladj(landscape)

# S3 method for list
lsm_c_pladj(landscape)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

Value

tibble

Details

$$PLADJ = (\frac{g_{ij}} {\sum \limits_{k = 1}^{m} g_{ik}}) * 100$$ where \(g_{ii}\) is the number of adjacencies between cells of class i and \(g_{ik}\) is the number of adjacencies between cells of class i and k.

PLADJ is an 'Aggregation metric'. It calculates the frequency how often patches of different classes i (focal class) and k are next to each other, and following is a measure of class aggregation. The adjacencies are counted using the double-count method.

Units
Percent
Ranges
0 <= PLADJ <= 100
Behaviour
Equals PLADJ = 0 if class i is maximal disaggregated, i.e. every cell is a different patch. Equals PLADJ = 100 when the only one patch is present.
References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html.

Examples

lsm_c_pladj(landscape)

#> [38;5;246m# A tibble: 3 x 6[39m
#>   layer level class    id metric value
#>   [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m  [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m     1 class     1    [31mNA[39m pladj   72.6
#> [38;5;250m2[39m     1 class     2    [31mNA[39m pladj   72.7
#> [38;5;250m3[39m     1 class     3    [31mNA[39m pladj   79.7