PLAND (class level)

Percentage of landscape of class (Area and Edge metric)

Usage

lsm_c_pland(landscape, directions = 8)

Arguments

landscape

A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters.

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$PLAND = \frac{\sum \limits_{j = 1}^{n} a_{ij}} {A} * 100$$ where \(a_{ij}\) is the area of each patch and \(A\) is the total landscape area.

PLAND is an 'Area and edge metric'. It is the percentage of the landscape belonging to class i. It is a measure of composition and because of the relative character directly comparable among landscapes with different total areas.

Because the metric is based on distances or areas please make sure your data is valid using check_landscape.

Units

Percentage

Range

0 < PLAND <= 100

Behaviour

Approaches PLAND = 0 when the proportional class area is decreasing. Equals PLAND = 100 when only one patch is present.

References

McGarigal K., SA Cushman, and E Ene. 2023. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical Maps. Computer software program produced by the authors; available at the following web site: https://www.fragstats.org

See also

Examples

landscape <- terra::rast(landscapemetrics::landscape)
lsm_c_pland(landscape)
#> # A tibble: 3 × 6
#>   layer level class    id metric value
#>   <int> <chr> <int> <int> <chr>  <dbl>
#> 1     1 class     1    NA pland   20.4
#> 2     1 class     2    NA pland   26  
#> 3     1 class     3    NA pland   53.6