FRAC_SD (landscape level)

Standard deviation fractal dimension index (Shape metric)

Usage

lsm_l_frac_sd(landscape, directions = 8)

Arguments

landscape

Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$FRAC_{SD} = sd(FRAC[patch_{ij}])$$ where \(FRAC[patch_{ij}]\) equals the fractal dimension index of each patch.

FRAC_SD is a 'Shape metric'. The metric summarises the landscape as the standard deviation of the fractal dimension index of all patches in the landscape. The fractal dimension index is based on the patch perimeter and the patch area and describes the patch complexity.

Units

None

Range

FRAC_SD>= 0

Behaviour

Equals FRAC_SD = 0 if the fractal dimension index is identical for all patches. Increases, without limit, as the variation of the fractal dimension indices increases.

References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Mandelbrot, B. B. 1977. Fractals: Form, Chance, and Dimension. San Francisco. W. H. Freeman and Company.

See also

Examples

lsm_l_frac_sd(landscape)
#> # A tibble: 1 × 6
#>   layer level     class    id metric  value
#>   <int> <chr>     <int> <int> <chr>   <dbl>
#> 1     1 landscape    NA    NA frac_sd 0.207