Interspersion and Juxtaposition index (landscape level)

Interspersion and Juxtaposition index (Aggregation metric)

Usage

lsm_l_iji(landscape, verbose = TRUE)

Arguments

landscape

A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters.

verbose

Print warning message if not sufficient patches are present

Value

tibble

Details

$$IJI = \frac{- \sum \limits_{i = 1}^{m} \sum \limits_{k = i + 1}^{m} \Bigg[ \Bigg( \frac{e_{ik}}{E} \Bigg) ln \Bigg( \frac{e_{ik}}{E} \Bigg) \Bigg]}{ln(0.5[m(m - 1)])} * 100$$

where \(e_{ik}\) are the unique adjacencies of all classes (lower/upper triangle of the adjacency table - without the diagonal), \(E\) is the total length of edges in the landscape and \(m\) is the number of classes.

IJI is an 'Aggregation metric'. It is a so called "salt and pepper" metric and describes the intermixing of classes (i.e. without considering like adjacencies - the diagonal of the adjacency table). The number of classes to calculate IJI must be >= than 3.

Units

Percent

Range

0 < IJI <= 100

Behaviour

Approaches 0 if a class is only adjacent to a single other class and equals 100 when a class is equally adjacent to all other classes.

References

McGarigal K., SA Cushman, and E Ene. 2023. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical Maps. Computer software program produced by the authors; available at the following web site: https://www.fragstats.org

McGarigal, K., & Marks, B. J. 1995. FRAGSTATS: spatial pattern analysis program for quantifying landscape structure. Gen. Tech. Rep. PNW-GTR-351. Portland, OR: US Department of Agriculture, Forest Service, Pacific Northwest Research Station. 122 p, 351.

See also

Examples

landscape <- terra::rast(landscapemetrics::landscape)
lsm_l_iji(landscape)
#> # A tibble: 1 × 6
#>   layer level     class    id metric value
#>   <int> <chr>     <int> <int> <chr>  <dbl>
#> 1     1 landscape    NA    NA iji     89.8