MSIDI (landscape level)

Modified Simpson's diversity index (Diversity metric)

lsm_l_msidi(landscape, directions)

# S3 method for RasterLayer
lsm_l_msidi(landscape, directions = 8)

# S3 method for RasterStack
lsm_l_msidi(landscape, directions = 8)

# S3 method for RasterBrick
lsm_l_msidi(landscape, directions = 8)

# S3 method for stars
lsm_l_msidi(landscape, directions = 8)

# S3 method for list
lsm_l_msidi(landscape, directions = 8)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

directions
The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$MSIDI = -\ln \sum \limits_{i = 1}^{m} P_{i}^{2}$$ where \(P_{i}\) is the landscape area proportion of class i.

MSIDI is a 'Diversity metric'.

Units
None
Range
MSIDI >= 0
Behaviour
MSIDI = 0 if only one patch is present and increases, without limit, as the amount of patches with equally distributed landscape proportions increases
References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Simpson, E. H. 1949. Measurement of diversity. Nature 163:688

Pielou, E. C. 1975. Ecological Diversity. Wiley-Interscience, New York.

Romme, W. H. 1982. Fire and landscapediversity in subalpine forests of Yellowstone National Park.Ecol.Monogr. 52:199-221

See also

Examples

lsm_l_msidi(landscape)

#> [38;5;246m# A tibble: 1 x 6[39m
#>   layer level     class    id metric value
#>   [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m  [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m     1 landscape    [31mNA[39m    [31mNA[39m msidi  0.929