Modified Simpson's evenness index (landscape level)

Modified Simpson's evenness index (landscape level)

lsm_l_msiei(landscape)

# S3 method for RasterLayer
lsm_l_msiei(landscape)

# S3 method for RasterStack
lsm_l_msiei(landscape)

# S3 method for RasterBrick
lsm_l_msiei(landscape)

# S3 method for list
lsm_l_msiei(landscape)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

Value

tibble

Details

The modified Simpson's evenness index equals the negative natural logarithm of the sum of the squared class proportions divided by the natural logarithm of number of classes $$MSIDI = -ln(sum(proportion[patch_i] ^ 2)) / ln(patch richness)$$

Units
None
Range
0 <= MSIEI < 1
Behaviour
MSIEI = 0 when only one class and patch is present and approaches MSIEI = 1 as the proportional distribution of patches becomes more even
References

McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep. PNW-351.

Examples

lsm_l_msiei(landscape)

#> # A tibble: 1 x 6
#>   layer level     class    id metric                            value
#>   <int> <chr>     <int> <int> <chr>                             <dbl>
#> 1     1 landscape    NA    NA modified Simpson's evenness index 0.845