NP (landscape level)

Number of patches (Aggregation metric)

lsm_l_np(landscape, directions = 8)

Arguments

landscape
Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

directions
The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$NP = N$$ where \(N\) is the number of patches.

NP is an 'Aggregation metric'. It describes the fragmentation of the landscape, however, does not necessarily contain information about the configuration or composition of the landscape.

Units

None

Ranges

NP >= 1

Behaviour

Equals NP = 1 when only one patch is present and increases, without limit, as the number of patches increases

References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

See also

Examples

lsm_l_np(landscape)

#> # A tibble: 1 x 6
#>   layer level     class    id metric value
#>   <int> <chr>     <int> <int> <chr>  <dbl>
#> 1     1 landscape    NA    NA np        27