PARA_MN (landscape level)

Mean perimeter-area ratio (Shape metric)

Usage

lsm_l_para_mn(landscape, directions = 8)

Arguments

landscape

A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters.

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$PARA_{MN} = mean(PARA[patch_{ij}]$$ where \(PARA[patch_{ij}]\) is the perimeter area ratio of each patch.

PARA_MN is a 'Shape metric'. It summarises the landscape as the mean of each patch in the landscape. The perimeter-area ratio describes the patch complexity in a straightforward way. However, because it is not standarised to a certain shape (e.g. a square), it is not scale independent, meaning that increasing the patch size while not changing the patch form will change the ratio. Because the metric is based on distances or areas please make sure your data is valid using check_landscape.

Units

None

Range

PARA_MN > 0

Behaviour

Approaches PARA_MN > 0 if PARA for each patch approaches PARA > 0, i.e. the form approaches a rather small square. Increases, without limit, as PARA increases, i.e. patches become more complex.

References

McGarigal K., SA Cushman, and E Ene. 2023. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical Maps. Computer software program produced by the authors; available at the following web site: https://www.fragstats.org

See also

Examples

landscape <- terra::rast(landscapemetrics::landscape)
lsm_l_para_mn(landscape)
#> # A tibble: 1 × 6
#>   layer level     class    id metric  value
#>   <int> <chr>     <int> <int> <chr>   <dbl>
#> 1     1 landscape    NA    NA para_mn  2.16