RELMUTINF (landscape level)

Relative mutual information

Usage

lsm_l_relmutinf(landscape, neighbourhood = 4, ordered = TRUE, base = "log2")

Arguments

landscape

A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters.

neighbourhood

The number of directions in which cell adjacencies are considered as neighbours: 4 (rook's case) or 8 (queen's case). The default is 4.

ordered

The type of pairs considered. Either ordered (TRUE) or unordered (FALSE). The default is TRUE.

base

The unit in which entropy is measured. The default is "log2", which compute entropy in "bits". "log" and "log10" can be also used.

Value

tibble

Details

Due to the spatial autocorrelation, the value of mutual information tends to grow with a diversity of the landscape (marginal entropy). To adjust this tendency, it is possible to calculate relative mutual information by dividing the mutual information by the marginal entropy. Relative mutual information always has a range between 0 and 1 and can be used to compare spatial data with different number and distribution of categories. When the value of mutual information equals to 0, then relative mutual information is 1.

References

Nowosad J., TF Stepinski. 2019. Information theory as a consistent framework for quantification and classification of landscape patterns. https://doi.org/10.1007/s10980-019-00830-x

See also

Examples

landscape <- terra::rast(landscapemetrics::landscape)
lsm_l_relmutinf(landscape)
#> # A tibble: 1 × 6
#>   layer level     class    id metric    value
#>   <int> <chr>     <int> <int> <chr>     <dbl>
#> 1     1 landscape    NA    NA relmutinf 0.400