SPLIT (landscape level)

Splitting index (Aggregation metric)

lsm_l_split(landscape, directions)

# S3 method for RasterLayer
lsm_l_split(landscape, directions = 8)

# S3 method for RasterStack
lsm_l_split(landscape, directions = 8)

# S3 method for RasterBrick
lsm_l_split(landscape, directions = 8)

# S3 method for list
lsm_l_split(landscape, directions = 8)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

directions
The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$SSPLIT = \frac{A^2} {\sum \limits_{i = 1}^{m} \sum \limits_{j = 1}^{n} a_{ij}^2}$$ where \(a_{ij}\) is the patch area in square meters and \(A\) is the total landscape area.

SPLIT is an 'Aggregation metric'. It describes the number of patches if all patches the landscape would be divided into equally sized patches.

Units
None
Range
1 <= SPLIT <= Number of cells squared
Behaviour
Equals SPLIT = 1 if only one patch is present. Increases as the number of patches increases and is limited if all cells are a patch
References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Jaeger, J. A. 2000. Landscape division, splitting index, and effective mesh size: new measures of landscape fragmentation. Landscape ecology, 15(2), 115-130.

See also

Examples

lsm_l_split(landscape)

#> # A tibble: 1 x 6
#>   layer level     class    id metric value
#>   <int> <chr>     <int> <int> <chr>  <dbl>
#> 1     1 landscape    NA    NA split   3.29