NCORE (patch level)

Number of core areas (Core area metric)

lsm_p_ncore(
  landscape,
  directions = 8,
  consider_boundary = FALSE,
  edge_depth = 1
)

Arguments

landscape
Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

directions
The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

consider_boundary
Logical if cells that only neighbour the landscape boundary should be considered as core

edge_depth
Distance (in cells) a cell has the be away from the patch edge to be considered as core cell

#' @details $$NCORE = n_{ij}^{core}$$ where \(n_{ij}^{core}\) is the number of disjunct core areas.

NCORE is a 'Core area metric'. A cell is defined as core if the cell has no neighbour with a different value than itself (rook's case). The metric counts the disjunct core areas, whereby a core area is a 'patch within the patch' containing only core cells. It describes patch area and shape simultaneously (more core area when the patch is large, however, the shape must allow disjunct core areas). Thereby, a compact shape (e.g. a square) will contain less disjunct core areas than a more irregular patch.

Units

None

Range

NCORE >= 0

Behaviour

NCORE = 0 when CORE = 0, i.e. every cell in patch is edge. Increases, without limit, as core area increases and patch shape allows disjunct core areas (i.e. patch shape becomes rather complex).

Value

tibble

References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

See also

Examples

lsm_p_ncore(landscape)

#> # A tibble: 27 x 6
#>    layer level class    id metric value
#>    <int> <chr> <int> <int> <chr>  <dbl>
#>  1     1 patch     1     1 ncore      0
#>  2     1 patch     1     2 ncore      0
#>  3     1 patch     1     3 ncore      3
#>  4     1 patch     1     4 ncore      0
#>  5     1 patch     1     5 ncore      0
#>  6     1 patch     1     6 ncore      2
#>  7     1 patch     1     7 ncore      0
#>  8     1 patch     1     8 ncore      0
#>  9     1 patch     1     9 ncore      0
#> 10     1 patch     2    10 ncore      1
#> # … with 17 more rows