PERIM (patch level)

Perimeter (Area and edge metric))

Usage

lsm_p_perim(landscape, directions = 8)

Arguments

landscape

Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$PERIM = p_{ij}$$ where \(p_{ij}\) is the perimeter in meters.

PERIM is an 'Area and edge metric'. It equals the perimeter of the patch including also the edge to the landscape boundary. The metric describes patch area (larger perimeter for larger patches), but also patch shape (large perimeter for irregular shapes).

Units

Meters

Range

PERIM > 0

Behaviour

Increases, without limit, as patch size and complexity increases.

References

McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical and Continuous Maps. Computer software program produced by the authors at the University of Massachusetts, Amherst. Available at the following web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Examples

lsm_p_perim(landscape)
#> # A tibble: 27 × 6
#>    layer level class    id metric value
#>    <int> <chr> <int> <int> <chr>  <dbl>
#>  1     1 patch     1     1 perim      4
#>  2     1 patch     1     2 perim     12
#>  3     1 patch     1     3 perim    130
#>  4     1 patch     1     4 perim      4
#>  5     1 patch     1     5 perim      4
#>  6     1 patch     1     6 perim     20
#>  7     1 patch     1     7 perim      8
#>  8     1 patch     1     8 perim     10
#>  9     1 patch     1     9 perim      4
#> 10     1 patch     2    10 perim     38
#> # … with 17 more rows