matrix_to_raster

Convert matrix to raster

Usage

matrix_to_raster(
  matrix,
  landscape = NULL,
  landscape_empty = FALSE,
  extent = NULL,
  resolution = NULL,
  crs = NULL,
  to_disk = getOption("to_disk", default = FALSE)
)

Arguments

matrix

matrix with values.

landscape

SpatRaster

landscape_empty

If true, SpatRaster is landscape_empty

extent

Extent of SpatRaster

resolution

Resolution of SpatRaster

crs

CRS of raster layer.

to_disk

If TRUE raster will be saved to disk.

Value

raster

Details

Converts matrix to a raster with same characteristics as landscape. Either landscape or extent, resolution and crs must be specified.

Examples

augusta_nlcd <- terra::rast(landscapemetrics::augusta_nlcd)

test_matrix <- terra::as.matrix(augusta_nlcd, wide = TRUE)
matrix_to_raster(matrix = test_matrix, landscape = augusta_nlcd)
#> class       : SpatRaster 
#> dimensions  : 440, 678, 1  (nrow, ncol, nlyr)
#> resolution  : 30, 30  (x, y)
#> extent      : 1249665, 1270005, 1246815, 1260015  (xmin, xmax, ymin, ymax)
#> coord. ref. : Albers Conical Equal Area 
#> source(s)   : memory
#> name        : lyr.1 
#> min value   :    11 
#> max value   :    95