extract_lsm

Extract metrics

extract_lsm(landscape, y, what, metric, name, type, directions,
  consider_boundary, edge_depth, full_name, verbose, progress, ...)

# S3 method for RasterLayer
extract_lsm(landscape, y, what = NULL,
  metric = NULL, name = NULL, type = NULL, directions = 8,
  consider_boundary = FALSE, edge_depth = 1, full_name = FALSE,
  verbose = TRUE, progress = FALSE, ...)

# S3 method for RasterStack
extract_lsm(landscape, y, what = NULL,
  metric = NULL, name = NULL, type = NULL, directions = 8,
  consider_boundary = FALSE, edge_depth = 1, full_name = FALSE,
  verbose = TRUE, progress = FALSE, ...)

# S3 method for RasterBrick
extract_lsm(landscape, y, what = NULL,
  metric = NULL, name = NULL, type = NULL, directions = 8,
  consider_boundary = FALSE, edge_depth = 1, full_name = FALSE,
  verbose = TRUE, progress = FALSE, ...)

# S3 method for stars
extract_lsm(landscape, y, what = NULL, metric = NULL,
  name = NULL, type = NULL, directions = 8,
  consider_boundary = FALSE, edge_depth = 1, full_name = FALSE,
  verbose = TRUE, progress = FALSE, ...)

# S3 method for list
extract_lsm(landscape, y, what = NULL, metric = NULL,
  name = NULL, type = NULL, directions = 8,
  consider_boundary = FALSE, edge_depth = 1, full_name = FALSE,
  verbose = TRUE, progress = FALSE, ...)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

y
Spatial object ( Spatialy*; SpatialPolygons*; SpatialLines; Extent or sf equivalents); two-column matrix/data.frame/tibble or cellnumbers that are used to extract landscapemetrics.

what
String indicating what metric to calculate, either "patch" (default) for all patch level metrics or any of the patch metrics functions.

metric
Abbreviation of metrics to calculate (e.g. 'area').

name
Full name of metrics to calculate (e.g. 'core area').

type
Metric types to calculate according to FRAGSTATS grouping (e.g. 'aggregation metric').

directions
The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

consider_boundary
Logical if cells that only neighbour the landscape boundary should be considered as core

edge_depth
Distance (in cells) a cell has the be away from the patch edge to be considered as core cell

full_name
Should the full names of all functions be included in the tibble.

verbose
Print warning message if not sufficient patches are present

progress
Print progress report

...
Arguments passed to raster::extract.

Value

tibble

Details

This functions extracts the metrics of all patches the spatial object(s) y (e.g. spatial points) are located within. Only patch level metrics are possible to extract. Please be aware that the output is sligthly different to all other lsm-function of landscapemetrics. Returns a tibble with chosen metrics and the ID of the spatial objects.

See also

Examples

points <- matrix(c(10, 5, 25, 15, 5, 25), ncol = 2, byrow = TRUE)
extract_lsm(landscape, points)

#> Warning: return_raster = TRUE not able for matrix input

#> Warning: return_raster = TRUE not able for matrix input

#> Warning: return_raster = TRUE not able for matrix input

#> Warning: return_raster = TRUE not able for matrix input

#> Warning: return_raster = TRUE not able for matrix input

#> Warning: return_raster = TRUE not able for matrix input

#> # A tibble: 36 x 7
#>    layer level class    id metric   value extract_id
#>    <chr> <chr> <int> <dbl> <chr>    <dbl>      <int>
#>  1 1     patch     3    25 area    0.0457          1
#>  2 1     patch     3    25 cai    57.5             1
#>  3 1     patch     3    25 circle  0.631           1
#>  4 1     patch     3    25 contig  0.785           1
#>  5 1     patch     3    25 core    0.0263          1
#>  6 1     patch     3    25 enn     2               1
#>  7 1     patch     3    25 frac    1.46            1
#>  8 1     patch     3    25 gyrate 11.2             1
#>  9 1     patch     3    25 ncore  11               1
#> 10 1     patch     3    25 para    0.761           1
#> # ... with 26 more rows

extract_lsm(landscape, points, type = "aggregation metric")

#> # A tibble: 3 x 7
#>   layer level class    id metric value extract_id
#>   <chr> <chr> <int> <dbl> <chr>  <dbl>      <int>
#> 1 1     patch     3    25 enn     2             1
#> 2 1     patch     3    25 enn     2             2
#> 3 1     patch     2    10 enn     4.47          3

points_sp <- sp::SpatialPoints(points)
extract_lsm(landscape, points, what = "lsm_p_area")

#> # A tibble: 3 x 7
#>   layer level class    id metric  value extract_id
#>   <chr> <chr> <int> <dbl> <chr>   <dbl>      <int>
#> 1 1     patch     3    25 area   0.0457          1
#> 2 1     patch     3    25 area   0.0457          2
#> 3 1     patch     2    10 area   0.0035          3