extract_lsm

Extract metrics

Usage

extract_lsm(
  landscape,
  y,
  extract_id = NULL,
  metric = NULL,
  name = NULL,
  type = NULL,
  what = NULL,
  directions = 8,
  progress = FALSE,
  verbose = TRUE,
  ...
)

Arguments

landscape

Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

y

2-column matrix with coordinates, SpatialPoints, SpatialLines or sf point geometries.

extract_id

Vector with id of sample points. If not provided, sample points will be labelled 1...n.

metric

Abbreviation of metrics (e.g. 'area').

name

Full name of metrics (e.g. 'core area')

type

Type according to FRAGSTATS grouping (e.g. 'aggregation metrics').

what

Selected level of metrics: either "patch", "class" or "landscape". It is also possible to specify functions as a vector of strings, e.g. what = c("lsm_c_ca", "lsm_l_ta").

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

progress

Print progress report.

verbose

Print warning messages.

...

Arguments passed to calculate_lsm().

Value

tibble

Details

This functions extracts the metrics of all patches the spatial object(s) y (e.g. spatial points) are located within. Only patch level metrics are possible to extract. Please be aware that the output is sligthly different to all other lsm-function of landscapemetrics. Returns a tibble with chosen metrics and the ID of the spatial objects.

See also

Examples

points <- matrix(c(10, 5, 25, 15, 5, 25), ncol = 2, byrow = TRUE)
extract_lsm(landscape, y = points)
#> Warning: Please use 'check_landscape()' to ensure the input data is valid.
#> # A tibble: 36 × 7
#>    layer level class    id metric   value extract_id
#>    <int> <chr> <int> <int> <chr>    <dbl>      <dbl>
#>  1     1 patch     3    24 area    0.0457          1
#>  2     1 patch     3    24 cai    49.2             1
#>  3     1 patch     3    24 circle  0.666           1
#>  4     1 patch     3    24 contig  0.785           1
#>  5     1 patch     3    24 core    0.0225          1
#>  6     1 patch     3    24 enn     2               1
#>  7     1 patch     3    24 frac    1.46            1
#>  8     1 patch     3    24 gyrate 11.2             1
#>  9     1 patch     3    24 ncore  11               1
#> 10     1 patch     3    24 para    0.761           1
#> # … with 26 more rows
extract_lsm(landscape, y = points, type = "aggregation metric")
#> Warning: Please use 'check_landscape()' to ensure the input data is valid.
#> # A tibble: 3 × 7
#>   layer level class    id metric value extract_id
#>   <int> <chr> <int> <int> <chr>  <dbl>      <dbl>
#> 1     1 patch     3    24 enn     2             1
#> 2     1 patch     3    24 enn     2             2
#> 3     1 patch     2    10 enn     4.47          3

points_sp <- sp::SpatialPoints(points)
extract_lsm(landscape, y = points_sp, what = "lsm_p_area")
#> Warning: Only using 'what' argument.
#> Warning: Please use 'check_landscape()' to ensure the input data is valid.
#> # A tibble: 3 × 7
#>   layer level class    id metric  value extract_id
#>   <int> <chr> <int> <int> <chr>   <dbl>      <dbl>
#> 1     1 patch     3    24 area   0.0457          1
#> 2     1 patch     3    24 area   0.0457          2
#> 3     1 patch     2    10 area   0.0035          3

if (FALSE) {
# use lines (works only if rgeos is installed)
x1 <- c(1, 5, 15, 10)
y1 <- c(1, 5, 15, 25)

x2 <- c(10, 25)
y2 <- c(5, 5)

sample_lines <- sp::SpatialLines(list(sp::Lines(list(sp::Line(cbind(x1, y1)),
sp::Line(cbind(x2, y2))), ID = "a")))
extract_lsm(landscape, y = sample_lines, what = "lsm_p_area")
}