get_adjacencies

Fast calculation of adjacencies between classes in a raster

get_adjacencies(landscape, neighbourhood = 4, what = "full",
  upper = FALSE)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

neighbourhood
The number of directions in which cell adjacencies are considered as neighbours: 4 (rook's case), 8 (queen's case) or a binary matrix where the ones define the neighbourhood. The default is 4.

what
Which adjacencies to calculate: "full" for a full adjacency matrix, "like" for the diagonal, "unlike" for the off diagonal part of the matrix and "triangle" for a triangular matrix counting adjacencies only once.

upper
Logical value indicating whether the upper triangle of the adjacency matrix should be returned (default FALSE).

Value

matrix

Details

A fast implementation with Rcpp to calculate the adjacency matrix for raster. The adjacency matrix is most often used in landscape metrics to describe the configuration of landscapes, is it is a cellwise count of edges between classes.

The "full" adjacency matrix is double-count method, as it contains the pairwise counts of cells between all classes. The diagonal of this matrix contains the like adjacencies, a count for how many edges a shared in each class with the same class. The "unlike" adjacencies are counting the cellwise edges between different classes.

Examples

# calculate full adjacency matrix
get_adjacencies(landscape, 4)

#>     1   2    3
#> 1 520  43  137
#> 2  43 704  184
#> 3 137 184 1528

# count diagonal neighbour adjacencies
diagonal_matrix <- matrix(c(1,  NA,  1,
                            NA,  0, NA,
                            1,  NA,  1), 3, 3, byrow = TRUE)
get_adjacencies(landscape, diagonal_matrix)

#>     1   2    3
#> 1 466  53  167
#> 2  53 602  240
#> 3 167 240 1376

# equivalent with the raster package:
adjacencies <- raster::adjacent(landscape, 1:raster::ncell(landscape), 4, pairs=TRUE)
table(landscape[adjacencies[,1]], landscape[adjacencies[,2]])

#>    
#>        1    2    3
#>   1  520   43  137
#>   2   43  704  184
#>   3  137  184 1528