get_centroids

Centroid of patches

Usage

get_centroids(
  landscape,
  directions = 8,
  cell_center = FALSE,
  return_sp = FALSE,
  verbose = TRUE
)

Arguments

landscape

Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

cell_center

If true, the coordinates of the centroid are forced to be a cell center within the patch.

return_sp

If true, a SpatialPointsDataFrame is returned.

verbose

Print warning messages

Details

Get the coordinates of the centroid of each patch. The centroid is by default defined as the mean location of all cell centers. To force the centroid to be located within each patch, use the cell_center argument. In this case, the centroid is defined as the cell center that is the closest to the mean location.

Examples

# get centroid location
get_centroids(landscape)
#> # A tibble: 27 × 6
#>    layer level class    id     x     y
#>    <int> <chr> <int> <int> <dbl> <dbl>
#>  1     1 patch     1     1  0.5  29.5 
#>  2     1 patch     1     2  0.7  20.9 
#>  3     1 patch     1     3 13.8  18.0 
#>  4     1 patch     1     4  0.5   2.5 
#>  5     1 patch     1     5  0.5   0.5 
#>  6     1 patch     1     6 16.9   8.21
#>  7     1 patch     1     7 20.5   0.5 
#>  8     1 patch     1     8 26.1   2.3 
#>  9     1 patch     1     9 29.5   6.5 
#> 10     1 patch     2    10  2.81 25.2 
#> # … with 17 more rows