Tibble of abbreviations coming from FRAGSTATS

A single tibble for every abbreviation of every metric that is reimplemented in landscapemetrics and its corrosponding full name in the literature.

lsm_abbreviations_names

Format

A tibble object.

Details

Can be used ofter calculating the metric(s) with a join to have a more readable results tibble or for visualizing your results.

Examples

aggregation_index <- lsm_p_area(landscape)
aggregation_index <- dplyr::right_join(aggregation_index,
                                       lsm_abbreviations_names,
                                       by = 'metric')
aggregation_index

#> # A tibble: 106 x 8
#>    layer level class    id metric    value full_name  type               
#>    <int> <chr> <int> <int> <chr>     <dbl> <chr>      <chr>              
#>  1     1 patch     1     1 area   0.0001   patch area area and ege metric
#>  2     1 patch     1     2 area   0.0148   patch area area and ege metric
#>  3     1 patch     1     3 area   0.0005   patch area area and ege metric
#>  4     1 patch     1     4 area   0.0014   patch area area and ege metric
#>  5     1 patch     1     5 area   0.0001   patch area area and ege metric
#>  6     1 patch     1     6 area   0.0005   patch area area and ege metric
#>  7     1 patch     1     7 area   0.0001   patch area area and ege metric
#>  8     1 patch     1     8 area   0.0001   patch area area and ege metric
#>  9     1 patch     1     9 area   0.000300 patch area area and ege metric
#> 10     1 patch     2    10 area   0.0035   patch area area and ege metric
#> # ... with 96 more rows