Tibble of abbreviations coming from FRAGSTATS

A single tibble for every abbreviation of every metric that is reimplemented in landscapemetrics and its corresponding full name in the literature.

lsm_abbreviations_names

Format

A tibble object.

Details

Can be used after calculating the metric(s) with a join to have a more readable results tibble or for visualizing your results.

Examples

aggregation_index <- lsm_p_area(landscape)

#> Error in get_patches_int(landscape, what = what, directions = directions) %>%     raster::as.list(): could not find function "%>%"

aggregation_index <- dplyr::right_join(aggregation_index,
                                       lsm_abbreviations_names,
                                       by = 'metric')

#> Error in dplyr::right_join(aggregation_index, lsm_abbreviations_names,     by = "metric"): object 'aggregation_index' not found

aggregation_index

#> Error in eval(expr, envir, enclos): object 'aggregation_index' not found