PERIM (patch level)

Perimeter (Area and edge metric)

Usage

lsm_p_perim(landscape, directions = 8)

Arguments

landscape

A categorical raster object: SpatRaster; Raster* Layer, Stack, Brick; stars or a list of SpatRasters.

directions

The number of directions in which patches should be connected: 4 (rook's case) or 8 (queen's case).

Value

tibble

Details

$$PERIM = p_{ij}$$ where \(p_{ij}\) is the perimeter in meters.

PERIM is an 'Area and edge metric'. It equals the perimeter of the patch including also the edge to the landscape boundary. The metric describes patch area (larger perimeter for larger patches), but also patch shape (large perimeter for irregular shapes).

Because the metric is based on distances or areas please make sure your data is valid using check_landscape.

Units

Meters

Range

PERIM > 0

Behaviour

Increases, without limit, as patch size and complexity increases.

References

McGarigal K., SA Cushman, and E Ene. 2023. FRAGSTATS v4: Spatial Pattern Analysis Program for Categorical Maps. Computer software program produced by the authors; available at the following web site: https://www.fragstats.org

Examples

landscape <- terra::rast(landscapemetrics::landscape)
lsm_p_perim(landscape)
#> # A tibble: 28 × 6
#>    layer level class    id metric value
#>    <int> <chr> <int> <int> <chr>  <dbl>
#>  1     1 patch     1     1 perim      4
#>  2     1 patch     1     2 perim     12
#>  3     1 patch     1     3 perim     60
#>  4     1 patch     1     4 perim      4
#>  5     1 patch     1     5 perim      4
#>  6     1 patch     1     6 perim     68
#>  7     1 patch     1     7 perim     22
#>  8     1 patch     1     8 perim     10
#>  9     1 patch     1     9 perim     10
#> 10     1 patch     2    10 perim     38
#> # ℹ 18 more rows