PERIM (patch level)

Perimeter (Area and edge metric))

lsm_p_perim(landscape)

# S3 method for RasterLayer
lsm_p_perim(landscape)

# S3 method for RasterStack
lsm_p_perim(landscape)

# S3 method for RasterBrick
lsm_p_perim(landscape)

# S3 method for list
lsm_p_perim(landscape)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

Value

tibble

Details

$$PERIM = p_{ij}$$ where \(p_{ij}\) is the perimeter in meters.

PERIM is an 'Area and edge metric'. It equals the perimeter of the patch including also the edge to the landscape boundary. The metric describes patch area (larger perimeter for larger patches), but also patch shape (large perimeter for irregular shapes).

Units
Meters
Range
PERIM > 0
Behaviour
Increases, without limit, as patch size and complexity increases.
References

McGarigal, K., and B. J. Marks. 1995. FRAGSTATS: spatial pattern analysis program for quantifying landscape structure. USDA For. Serv. Gen. Tech. Rep. PNW-351.

Examples

lsm_p_perim(landscape)

#> # A tibble: 27 x 6
#>    layer level class    id metric    value
#>    <int> <chr> <int> <int> <chr>     <dbl>
#>  1     1 patch     1     1 perimeter     4
#>  2     1 patch     1     2 perimeter    12
#>  3     1 patch     1     3 perimeter   130
#>  4     1 patch     1     4 perimeter     4
#>  5     1 patch     1     5 perimeter     4
#>  6     1 patch     1     6 perimeter    20
#>  7     1 patch     1     7 perimeter     8
#>  8     1 patch     1     8 perimeter    10
#>  9     1 patch     1     9 perimeter     4
#> 10     1 patch     2    10 perimeter    38
#> # ... with 17 more rows