sample_lsm

Sample metrics

sample_lsm(landscape, what, shape, points, size, return_plots, ...)

# S3 method for RasterLayer
sample_lsm(landscape, what = "all",
  shape = "square", points, size, return_plots = FALSE, ...)

# S3 method for RasterStack
sample_lsm(landscape, what = "all",
  shape = "square", points, size, return_plots = FALSE, ...)

# S3 method for RasterBrick
sample_lsm(landscape, what = "all",
  shape = "square", points, size, return_plots = FALSE, ...)

# S3 method for list
sample_lsm(landscape, what = "all", shape = "square",
  points, size, return_plots = FALSE, ...)

Arguments

landscape
Raster* Layer, Stack, Brick or a list of rasterLayers.

what
Selected level of metrics: either "all", "patch", "class", "landscape". The default is "all". It is also possible to specify functions as a vector of strings, e.g. what = c("lsm_c_ca", "lsm_l_ta").

shape
String specifying plot shape. Either "circle" or "square"

points
SpatialPoints, sf or 2-column matrix with coordinates of sample points

size
Approximated size of sample plot. Equals the radius for circles or the side-length for squares in mapunits

return_plots
Logical if the clipped raster of the sample plot should be returned

...
Options for calculate_lsm()

Value

tibble

Details

This function samples the selected metrics in a buffer area (sample plot) around sample points. The size of the acutal sampled landscape can be different to the provided size due to two reasons. Firstly, because clipping raster cells using a circle or a sample plot not directly at a cell center lead to inaccuracies. Secondly, sample plots can exceed the landscape boundary. Therefore, we report the actual clipped sample plot area relative in relation to the theoretical, maximum sample plot area e.g. a sample plot only half within the landscape will have a percentage_inside = 50. Please be aware that the output is sligthly different to all other lsm-function of landscapemetrics.

Examples

points <- matrix(c(10, 5, 25, 15, 5, 25), ncol = 2, byrow = TRUE)
sample_lsm(landscape, points = points, size = 15, what = "lsm_l_np")

#> [38;5;246m# A tibble: 3 x 8[39m
#>   layer plot_id level     class    id metric value percentage_inside
#>   [3m[38;5;246m<int>[39m[23m   [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m  [3m[38;5;246m<dbl>[39m[23m             [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m     1       1 landscape    [31mNA[39m    [31mNA[39m np         8              85.3
#> [38;5;250m2[39m     1       2 landscape    [31mNA[39m    [31mNA[39m np         7              74.7
#> [38;5;250m3[39m     1       3 landscape    [31mNA[39m    [31mNA[39m np         7              64. 

points_sp <- sp::SpatialPoints(points)
sample_lsm(landscape, points = points_sp, size = 15, what = "lsm_l_np", return_plots = TRUE)

#> [38;5;246m# A tibble: 3 x 4[39m
#>   layer plot_id metrics          raster_sample_plots
#>   [3m[38;5;246m<int>[39m[23m   [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<list>[39m[23m           [3m[38;5;246m<list>[39m[23m             
#> [38;5;250m1[39m     1       1 [38;5;246m<tibble [1 × 8]>[39m [38;5;246m<S4: RasterLayer>[39m  
#> [38;5;250m2[39m     1       2 [38;5;246m<tibble [1 × 8]>[39m [38;5;246m<S4: RasterLayer>[39m  
#> [38;5;250m3[39m     1       3 [38;5;246m<tibble [1 × 8]>[39m [38;5;246m<S4: RasterLayer>[39m