scale_sample

Metrics on changing sample scale

scale_sample(
  landscape,
  y,
  shape = "square",
  size,
  max_size,
  verbose = TRUE,
  progress = FALSE,
  ...
)

Arguments

landscape
Raster* Layer, Stack, Brick, SpatRaster (terra), stars, or a list of rasterLayers.

y
2-column matrix with coordinates or SpatialPoints.

shape
String specifying plot shape. Either "circle" or "square"

size
Approximated size of sample plot. Equals the radius for circles or half of the side-length for squares in mapunits. For lines size equals the width of the buffer.

max_size
Maximum size to which sample plot size is summed up.

verbose
Print warning messages.

progress
Print progress report.

...
Arguments passed on to calculate_lsm().

Value

tibble

Details

This function calculates the selected metrics in subsequential buffers around a/multiple point(s) of interest.

The size of the actual sampled landscape can be different to the provided size due to two reasons. Firstly, because clipping raster cells using a circle or a sample plot not directly at a cell center lead to inaccuracies. Secondly, sample plots can exceed the landscape boundary. Therefore, we report the actual clipped sample plot area relative in relation to the theoretical, maximum sample plot area e.g. a sample plot only half within the landscape will have a percentage_inside = 50. Please be aware that the output is sligthly different to all other lsm-function of landscapemetrics.

The metrics can be specified by the arguments what, level, metric, name and/or type (combinations of different arguments are possible (e.g. level = "class", type = "aggregation metric"). If an argument is not provided, automatically all possibilities are selected. Therefore, to get all available metrics, don't specify any of the above arguments.

See also

Examples

my_points = matrix(c(1265000, 1250000, 1255000, 1257000),
                  ncol = 2, byrow = TRUE)

scale_sample(landscape = augusta_nlcd, y = my_points,
size = 500, max_size = 5000, what = c("lsm_l_ent", "lsm_l_mutinf"))

#> Warning: The 'perecentage_inside' is below 90% for at least one buffer.

#> # A tibble: 40 x 9
#>    layer level     class    id metric value  size plot_id percentage_inside
#>    <int> <chr>     <int> <int> <chr>  <dbl> <dbl>   <int>             <dbl>
#>  1     1 landscape    NA    NA ent     2.27   500       1              98.0
#>  2     1 landscape    NA    NA ent     2.70  1000       1             101. 
#>  3     1 landscape    NA    NA ent     2.70  1500       1             100  
#>  4     1 landscape    NA    NA ent     2.82  2000       1              99.5
#>  5     1 landscape    NA    NA ent     2.93  2500       1             100. 
#>  6     1 landscape    NA    NA ent     2.97  3000       1             100  
#>  7     1 landscape    NA    NA ent     2.97  3500       1              95.4
#>  8     1 landscape    NA    NA ent     2.94  4000       1              90.1
#>  9     1 landscape    NA    NA ent     2.94  4500       1              85.3
#> 10     1 landscape    NA    NA ent     2.95  5000       1              81.8
#> # … with 30 more rows